Warning fixes (#66)
* bug: truncate on windows was not using length. Fixed it.
* warnings: fix unused parameter warnings when built w/o MPI support.
* warnings: fix typos in docs.
* warnings: fix shadowed declarations.
* warnings: implicit conversion between int and size_t.
* warnings: implicit conversion between `size_t` and `streamsize`.
* bug: fix accidental recursion in io::utils.
* warning: conversion between `int` and `size_t`.
* warnings: fix signedness mismatch
* warnings: unused parameter warning
* warnings: implicit conversion from `size_t` to `off_t`
* warnings: conversion between `size_t` and `long`
* warnings: fix size_t to ssize_t conversion warnings.
Also `write` returns a signed value, hence storing it as size_t is not
reasonable.
* warnings: int to size_t conversions.
* warnings: unused parameter
* warnings: unused parameters.
* warnings: implicit conversion from int to unsigned int.
* warnings: type conversion
* warnings: shadowed declaration.
* warnings: unused parameter warnings
* warnings: fix shadowed variable warning.
* warnings: shadowed variable warnings.

+++ include/diy/collection.hpp
@@ -40,9 +40,9 @@ namespace diy
       inline void   clear();
 
       int           add(Element e)                  { elements_.push_back(e); external_.push_back(-1); ++(*in_memory_.access()); return static_cast<int>(elements_.size()) - 1; }
-      void*         release(int i)                  { void* e = get(i); elements_[i] = 0; return e; }
+      void*         release(int i)                  { void* e = get(i); elements_[static_cast<size_t>(i)] = 0; return e; }
 
-      void*         find(int i) const               { return elements_[i]; }                        // possibly returns 0, if the element is unloaded
+      void*         find(int i) const               { return elements_[static_cast<size_t>(i)]; }                        // possibly returns 0, if the element is unloaded
       void*         get(int i)                      { if (!find(i)) load(i); return find(i); }      // loads the element first, and then returns its address
 
       int           available() const               { int i = 0; for (; i < (int)size(); ++i) if (find(i) != 0) break; return i; }
@@ -56,7 +56,7 @@ namespace diy
       Save          saver() const                   { return save_; }
 
       void*         create() const                  { return create_(); }
-      void          destroy(int i)                  { if (find(i)) { destroy_(find(i)); elements_[i] = 0; } else if (external_[i] != -1) storage_->destroy(external_[i]); }
+      void          destroy(int i)                  { if (find(i)) { destroy_(find(i)); elements_[static_cast<size_t>(i)] = 0; } else if (external_[static_cast<size_t>(i)] != -1) storage_->destroy(external_[static_cast<size_t>(i)]); }
 
       bool          own() const                     { return destroy_ != 0; }
 
@@ -95,10 +95,10 @@ unload(int i)
   void* e = find(i);
   //save_(e, bb);
   //external_[i] = storage_->put(bb);
-  external_[i] = storage_->put(e, save_);
+  external_[static_cast<size_t>(i)] = storage_->put(e, save_);
 
   destroy_(e);
-  elements_[i] = 0;
+  elements_[static_cast<size_t>(i)] = 0;
 
   --(*in_memory_.access());
 }
@@ -111,9 +111,9 @@ load(int i)
   //storage_->get(external_[i], bb);
   void* e = create_();
   //load_(e, bb);
-  storage_->get(external_[i], e, load_);
-  elements_[i] = e;
-  external_[i] = -1;
+  storage_->get(external_[static_cast<size_t>(i)], e, load_);
+  elements_[static_cast<size_t>(i)] = e;
+  external_[static_cast<size_t>(i)] = -1;
 
   ++(*in_memory_.access());
 }
